Fix small namespace issue (#2785)
Move folders
