+++ Resources/Locale/en-US/_NF/storage/components/magnet-pickup-components.ftl
